Remove a dependency of the SPMTestSupport target on the Command target, which in turn depends on Build, etc.  This lets the subset of unit tests that doesn't actually require the command line tools to still be used to test libSwiftPMDataModel.
The only real code change needed is to move two tests that are actually testing handling of command line arguments of the `swift-build` tool from BuildPlanTests to BuildToolTests.
