+++ Tests/BuildTests/BuildPlanTests.swift
@@ -2572,22 +2572,6 @@ final class BuildPlanTests: XCTestCase {
         try testBinaryTargets(platform: "macos", arch: "arm64e", destinationTriple: arm64eTriple)
     }
 
-    func testCreatingSanitizers() throws {
-        for sanitizer in Sanitizer.allCases {
-            XCTAssertEqual(sanitizer, try Sanitizer(argument: sanitizer.shortName))
-        }
-    }
-
-    func testInvalidSanitizer() throws {
-        do {
-            _ = try Sanitizer(argument: "invalid")
-            XCTFail("Should have failed to create Sanitizer")
-        } catch let error as ArgumentConversionError {
-            XCTAssertEqual(
-                error.description, "valid sanitizers: address, thread, undefined, scudo")
-        }
-    }
-
     func testAddressSanitizer() throws {
         try sanitizerTest(.address, expectedName: "address")
     }
